WID-130. Review and finalize venv workflow.
