+++ requirements.txt
@@ -14,6 +14,7 @@ nbformat
 requests
 gitpython
 pygithub
+pyunicore
 siibra
 tqdm
 tvb-widgets>=1.0
