+++ src/index.tsx
@@ -285,7 +285,6 @@ const xircuits: JupyterFrontEndPlugin<void> = {
         code_str += `command_str= "${command_str}"\n`;
         code_str += "p=Popen(command_str, stdout=PIPE, stderr=PIPE, universal_newlines=True, shell=True)\n";
         code_str += "print('Remote Execution in process...\\n')\n";
-        code_str += `print('Please go to ${config['url']} for more details\\n')\n`;
         code_str += `print('${config['msg']}\\n')\n`;
         code_str += "for line in p.stdout:\n";
         code_str += "    " + "print(line.rstrip())\n\n";
